Move PersistentWallet from linera-service to linera-core
Add fs feature to linera-core gating the linera-persistent and anyhow
dependencies. Move wallet persistence logic to
linera-core::environment::wallet::persistent as PersistentWallet,
re-exported from linera-service as Wallet. Display logic (pretty_print,
ChainDetails) stays in linera-service behind a WalletExt trait.

Fix thiserror_context type ambiguity in client_context.rs by explicitly
mapping IO errors to error::Inner::Io before wrapping with context,
needed now that linera-persistent is in the dependency graph.
