+++ linera-client/src/client_context.rs
@@ -816,9 +816,11 @@ impl<Env: Environment> ClientContext<Env> {
         info!("Loading bytecode files");
         let contract_bytecode = Bytecode::load_from_file(&contract)
             .await
+            .map_err(error::Inner::Io)
             .with_context(|| format!("failed to load contract bytecode from {:?}", &contract))?;
         let service_bytecode = Bytecode::load_from_file(&service)
             .await
+            .map_err(error::Inner::Io)
             .with_context(|| format!("failed to load service bytecode from {:?}", &service))?;
 
         info!("Publishing module");
@@ -850,10 +852,12 @@ impl<Env: Environment> ClientContext<Env> {
         blob_path: PathBuf,
     ) -> Result<CryptoHash, Error> {
         info!("Loading data blob file");
-        let blob_bytes = fs::read(&blob_path).context(format!(
-            "failed to load data blob bytes from {:?}",
-            &blob_path
-        ))?;
+        let blob_bytes = fs::read(&blob_path)
+            .map_err(error::Inner::Io)
+            .context(format!(
+                "failed to load data blob bytes from {:?}",
+                &blob_path
+            ))?;
 
         info!("Publishing data blob");
         self.apply_client_command(chain_client, |chain_client| {
