+++ linera-service/Cargo.toml
@@ -99,7 +99,7 @@ linera-base.workspace = true
 linera-bridge.workspace = true
 linera-chain.workspace = true
 linera-client = { workspace = true, features = ["fs"] }
-linera-core.workspace = true
+linera-core = { workspace = true, features = ["fs"] }
 linera-execution = { workspace = true, features = ["fs"] }
 linera-faucet-client.workspace = true
 linera-faucet-server.workspace = true
